refactor: split config/schema.ts into domain-specific schema files
